mediator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synm-mediator"
version = "0.1.0"
description = "Synm Personal AI Vault Mediator Service"
authors = [{name = "ky13"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "chromadb>=0.4.0",
    "sqlmodel>=0.0.14",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP"]